check_ISO_code {rWBclimate}	R Documentation

check country codes

Description

Checks if the country code entered is a valid country code that data exists for

Usage

check_ISO_code(iso)

Arguments

iso
The 3 letter country code based on ISO3 Country abbreviations (http://unstats.un.org/unsd/methods/m49/m49alpha.htm)

Value

TRUE if a valid code, otherwise an error is returned

Examples

## Not run: 
check_ISO_code("USA")

## End(Not run)

[Package rWBclimate version 0.1.5.91 Index]