check_locator {rWBclimate}	R Documentation

Checks for what kind of locator a user input

Description

Checks for what kind of locator a user input

Usage

check_locator(locator)

Arguments

locator
The ISO3 country code that you want data about. (http://unstats.un.org/unsd/methods/m49/m49alpha.htm) or the basin ID [1-468]

Value

geo_ref a string indicating what kind of geography to use in the api

[Package rWBclimate version 0.1.5.91 Index]