create_map_df {rWBclimate}	R Documentation

Create mapable dataframe

Description

A function that will download maps for a vector of basins or country codes and return a data frame that has the kml output processed such that it can be plotted with ggplot2 and other mapping functions:

Usage

create_map_df(locator)

Arguments

locator
The a vector of ISO3 country code's that you want data about. (http://unstats.un.org/unsd/methods/m49/m49alpha.htm) or the basin ID's [1-468] (http://data.worldbank.org/sites/default/files/climate_data_api_basins.pdf)

Details

kml files can be quite large (100k-600k per country) making downloading them every time you want to make a map time consuming. To reduce this time it's easiest to download kml files and store them. To set the directory use a line like this: options(kmlpath="/Users/emh/kmltemp") The option must be called "kmlpath". These files will be persistent until you delete them.

Examples

## Not run: 
to_map <- create_map_df(c("USA","MEX","CAN"))
ggplot(to_map, aes(x=long, y=lat,group=group))+ geom_polygon()

## End(Not run)

[Package rWBclimate version 0.1.5.91 Index]