date_correct {rWBclimate}	R Documentation

correct data values

Description

Round start and end dates to conform with data api standards. See api documentation (http://data.worldbank.org/developers/climate-data-api) for full details

Usage

date_correct(start, end)

Arguments

start
The start year

end
The end year

Value

a 2xM matrix where M in the number of periods in the data api

Examples

## Not run: 
date_correct(1921,1957)

## End(Not run)

[Package rWBclimate version 0.1.5.91 Index]