download_kml {rWBclimate}	R Documentation

Download kml files

Description

Downloads map data from in kml format and writes it to a temporary directory. You must specify a temporary directory to write files to in your options.

Usage

download_kml(locator)

Arguments

locator
The a vector of ISO3 country code's that you want data about. (http://unstats.un.org/unsd/methods/m49/m49alpha.htm) or the basin ID's [1-468] (http://data.worldbank.org/sites/default/files/climate_data_api_basins.pdf)

Details

kml files can be quite large making downloading them every time you want to make a map time consuming. To reduce this time it's easiest to download kml files and store them. To set the directory use a line like this: options(kmlpath="/Users/emh/kmltemp") The option must be called "kmlpath".

[Package rWBclimate version 0.1.5.91 Index]