get_climate_data {rWBclimate}	R Documentation

get_climate_data

Description

Download monthly average climate data from the world bank climate data api. Ideally you'll want to use the wrapper functions that call this.

Usage

get_climate_data(locator, geo_type, type, cvar, start, end)

Arguments

locator
The ISO3 country code that you want data about. (http://unstats.un.org/unsd/methods/m49/m49alpha.htm) or the basin ID [1-468]

geo_type
basin or country depending on the locator type

type
the type of data you want "mavg" for monthly averages, "annualavg"

cvar
The variable you're interested in. "pr" for precipitation, "tas" for temperature in celcius.

start
The starting year you want data for, can be in the past or the future. Must conform to the periods outlined in the world bank API. If given values don't conform to dates, the fuction will automatically round them.

end
The ending year you want data for, can be in the past or the future. Similar to the start date, dates will be rounded to the nearest end dat.

[Package rWBclimate version 0.1.5.91 Index]