get_ensemble_data_recursive {rWBclimate}	R Documentation

Wrapper for get_ensemble_climate_data()

Description

Function to recursively call the get_ensemble_climate_data(). Handles a vector of basins or countries as well as multiple dates.

Usage

get_ensemble_data_recursive(locator, geo_type, type, cvar, start, end)

Arguments

locator
The ISO3 country code that you want data about. (http://unstats.un.org/unsd/methods/m49/m49alpha.htm) or the basin ID [1-468]

geo_type
basin or country depending on the locator type

type
the type of data you want "mavg" for monthly averages, "annualavg"

cvar
The variable you're interested in. "pr" for precipitation, "tas" for temperature in celcius.

start
The starting year you want data for, can be in the past or the future. Must conform to the periods outlined in the world bank API. If given values don't conform to dates, the fuction will automatically round them.

end
The ending year you want data for, can be in the past or the future. Similar to the start date, dates will be rounded to the nearest end dat.

Examples

## Not run: 
 get_ensemble_data_recursive(c("1","2"),"basin","mavg","pr",1920,1940)

## End(Not run)

[Package rWBclimate version 0.1.5.91 Index]