get_historical_data {rWBclimate}	R Documentation

Download historical climate data

Description

The Climate Data API provides access to historical temperature and precipitation data. These data are separate from the outputs of the GCMs, and they are based on gridded climatologies from the Climate Research Unit.

Usage

get_historical_data(locator, cvar, time_scale)

Arguments

locator
The ISO3 country code that you want data about. (http://unstats.un.org/unsd/methods/m49/m49alpha.htm) or the basin ID [1-468]. The historical period for country is 1901 - 2009, and 1960 - 2009 for basin

cvar
The climate variable you're interested in. "pr" for precipitation, "tas" for temperature in celcius.

time_scale
The time scale you want to return values on. Must be "month", "year" or "decade"

Details

The time_scale parameter returns a different number of variables depending on the input timescale. Month will return 12 values, a historical average for that month across all years. Year will return yearly averages for each year, and decade will return decade averages.

Value

a dataframe with historical climate data

[Package rWBclimate version 0.1.5.91 Index]